Added Dropout to nn.functional
Removed dropout from precision.py
