+++ syft/frameworks/torch/nn/__init__.py
@@ -8,3 +8,4 @@ from syft.frameworks.torch.nn.rnn import LSTMCell
 from syft.frameworks.torch.nn.conv import Conv2d
 
 from syft.frameworks.torch.nn.pool import AvgPool2d
+from syft.frameworks.torch.nn.functional import dropout
